correction to task 0

+++ 0x12-singly_linked_lists/0-main.c
@@ -13,7 +13,7 @@ int main(void)
     list_t *head;
     list_t *new;
     list_t hello = {"World", 5, NULL};
-    size_t n;
+    unsigned int n;
 
     head = &hello;
     new = malloc(sizeof(list_t));
@@ -27,13 +27,13 @@ int main(void)
     new->next = head;
     head = new;
     n = print_list(head);
-    printf("-> %lu elements\n", n);
+    printf("-> %u elements\n", n);
 
     printf("\n");
     free(new->str);
     new->str = NULL;
     n = print_list(head);
-    printf("-> %lu elements\n", n);
+    printf("-> %u elements\n", n);
 
     free(new);
     return (0);
